+++ src/pudl/transform/ferc714.py
@@ -234,7 +234,7 @@ RENAME_COLS = {
         "respondent_id": "respondent_id_ferc714",
         "respondent_name": "respondent_name_ferc714",
     },
-    "core_ferc714__hourly_demand_pa": {
+    "core_ferc714__hourly_demand_by_planning_area": {
         "report_yr": "report_year",
         "plan_date": "report_date",
         "respondent_id": "respondent_id_ferc714",
@@ -397,7 +397,7 @@ def core_ferc714__respondent_id(
 
 
 @asset(io_manager_key="pudl_sqlite_io_manager")
-def core_ferc714__hourly_demand_pa(
+def core_ferc714__hourly_demand_by_planning_area(
     raw_ferc714__demand_hourly_pa: pd.DataFrame,
 ) -> pd.DataFrame:
     """Transform the hourly demand time series by Planning Area.
@@ -420,7 +420,8 @@ def core_ferc714__hourly_demand_pa(
     """
     logger.info("Converting dates into pandas Datetime types.")
     df = _pre_process(
-        raw_ferc714__demand_hourly_pa, table_name="core_ferc714__hourly_demand_pa"
+        raw_ferc714__demand_hourly_pa,
+        table_name="core_ferc714__hourly_demand_by_planning_area",
     )
 
     # Parse date strings
@@ -520,5 +521,7 @@ def core_ferc714__hourly_demand_pa(
         "demand_mwh",
     ]
     df = df.drop(columns=set(df.columns) - set(columns))
-    df = _post_process(df[columns], table_name="core_ferc714__hourly_demand_pa")
+    df = _post_process(
+        df[columns], table_name="core_ferc714__hourly_demand_by_planning_area"
+    )
     return df